Remove the additional margin around body
By default in the Chrome browser the body has a margin of ~8px. This
gets added "on top" of the existing width property of 100vw which means
that no matter what window size there is, the window will always be
slightly larger (2x8px) than the screen size.

This commit addresses this by replacing the margin with padding and
using the "box-sizing: border-box" to indicate to the browser the bounds
should be calculated on the box - padding.

This addresses the overflow issue.