Initial implementation of CI/CD
Currently this site is deployed on Netlify. However, there are issues
with the setup Netlify provides in CI/CD.

Specifically, the goal of this system is to use the non-standard "hugo"
static site generator with "Asciidoctor". This allows using the
feature-complete asciidoctor markup, but with the hugo site generator
for content graphing.

The definitions for CI/CD, the container and the deployment were all
duplicated from docs.littleman.co, which this will eventually replace.