Add the blog top level element
The goal is to have all littleman.co information in a central spot, and
to have that content be easily navigable. This commit goes some way to
do this by adding a top level heading for "blog".

Future work should consider whether "blog" is an appropriate top level
entity. Currently there is both "blog" and "docs", which are both large
literary sets of work. Additionally, there's "case studies" and
potentially `${N}` different types of long content in future, as well as
videos and so fourth.