Specify a specific build container version
There is a bug in the upstream hugo container that was fixed recently on
the master branch. It was earlier presumed that :latest would force a
"re-pull" o fthe image, but the image appears to be cached somewhere in
an intermediary.

This commit addresses this by using a tagged version of that image,
which should cache bust.