Remove superfleus blue border
In previous work the page was separated into two components:

1. The canonical article content
2. Additional metadata, such as the ability to raise an issue

However, there is a class collision with the word "content". This word
is also used the `code` block, and the blue line thus appears in
multiple places where it should not.

This commit addresses this by using the word "article" instead of
"content".