Update clarification request to allow articles
The difference between articles and documentation is the point in time
nature of articles provides some notion as to the freshness of that
content, as well as provides a narrative through that notion that
documentation may not.

However, in principle both documentation and articles are a request for
the same sort of conceptual knowledge graph.

Accordingly this issue is amended to request articles or videos as well
as documentation