Allow users to request articles
Given that the website is designed to help users understand software
engineering and the inherent complexities therein, users should be able
to make requests of articles they'd like to understand a given topic.

This commit provides this ability by cross-linking the website to the
"new issue" article type, which encourages users to lodge a request for
an article they'd like to read more about.