Increase size of headings
Currently there is a problem in which the headings are not super easily
deimitable. This could be addressed by changing the size or font of the
headings, but a larger heading looks OK.

The heading sizes were derived from:

  https://material.io/design/typography/the-type-system.html#

But were offset a single level as the headings were simply too large as
defined.