Encourage users to use Google to search the documentation
While it would be possible to create a search implementation for the
website including form fields and so fourth, such a search
implementation is unlikely to be the canonical means for accessing the
website.

Rather, the website will be indexed by Google et. al. It thus makes
sense to concentrate on optimizing for Google searchability rather than
creating our own search tool, and encourage users to use Google to query
the site.

This commit does so by linking users to Google if they wish to search
with the filter `site:www.littleman.co` prefilled.