Add styleguide for admonitions and references
Currently there is some confusion as to when to use a reference versus
simply linking to the content within an article.

This commit attempts to clarify that by providing guidelines as to when
to use a reference (essentially always), leaving links for other types
of content.

It additionally adds some guidelines for admonitions which should also
serve as a demonstration of the admonition design.

Resolves #23