Add a team section, andrew howden to it
Currently there is nothing that describes who is involved with the
organisation. Further, there is a redundant domain at "andrewhowden.com"
which is falling in to disrepair.

Rather than have both, this commit attempts to deduplicate much of the
responsibility of andrewhowden.com to be only at "littleman.co". While
this means that I'm fairly committed to consulting if I fail this
employment I think that's fine.