Deploy current image, block deployments
Currently there is a problem in which, while a deployment will execute
and an image should be built, it is not the most image deployed into the
production environment.

The hypothesis is that when deployments are executed frequently there is
nothing in place to "force" the update of the image to a specific hash.
Accordingly, the image is just pulled at the time its created.

When multiple deployments are run simultaneously an old pod and a new
pod might be spawned simultaneously, but are both still "valid" in the
existing design.

This commit addresses this in two ways:

1. Deploy the specific version of a pod build in the release cycle
2. Block the build from completing until the pods are up

This should prevent multiple releases from deploying the "Same" pod.