Add an example to the Goalng styleguide
The author is settling on "go" as their preferred language to write
software in, including business facing software.

Given this, the guidelines will be considerably expanded prior to adding
content to the `github.com/littlemanco/oilr--proj` project generator.

This commit facilitates this by adding an example on top of which later
documentation will build for these guidelines.