Update the canonical packages
With more experience and recent work writing go programs the author has
come to the use of the new OpenTelemetry libraries.

These libraries aim to solve the metrics and tracing component, as well
as at some ill-defined point in future the logging component.

Given this, it stands to reason that software should start to be written
with this set of components now, even while it is in development. While
it might be a little frustrating to have to continue to rework APIs, the
likelyhood is the author simply will not write that much software
between now and the release of these libraries.

Additionally, in playground software the author has already written such
tools and found them to be useful.

== Deprecating Logarus

The Logarus library is essentially a "structured event" generator. The
representation of event as "spans" is a better concept; given this, this
commit drops logarus for that tracing implemntation.

There are still some problems (such as the bootstrap of the telemetry
libraries themselves) that require logging, so the basic go log library
is kept as it is simple, and part of stdlib.